refactor: move penas listing slice into core application
