+++ backend/src/core/application/models/__init__.py
@@ -1,3 +1,4 @@
+from .pena_listing_models import PenaInfo, PenasPage, PenasPageResult, PenaSummary
 from .season_match_insights_models import (
     MatchDetail,
     MatchInsightRow,
@@ -10,4 +11,8 @@ __all__ = [
     "MatchInsightRow",
     "MatchPlayerStats",
     "MatchTeam",
+    "PenaInfo",
+    "PenasPage",
+    "PenasPageResult",
+    "PenaSummary",
 ]
