+++ backend/src/core/application/ports/__init__.py
@@ -1,3 +1,4 @@
+from .pena_query_port import PenaQueryPort
 from .season_match_insights_port import SeasonMatchInsightsPort
 
-__all__ = ["SeasonMatchInsightsPort"]
+__all__ = ["PenaQueryPort", "SeasonMatchInsightsPort"]
