+++ backend/src/core/application/use_cases/__init__.py
@@ -1,3 +1,4 @@
+from .get_penas_usecase import GetPenasUseCase
 from .get_season_match_insights_usecase import GetSeasonMatchInsightsUseCase
 from .season_match_insights_errors import (
     InvalidSeasonInsightsDataError,
@@ -6,6 +7,7 @@ from .season_match_insights_errors import (
 )
 
 __all__ = [
+    "GetPenasUseCase",
     "GetSeasonMatchInsightsUseCase",
     "InvalidSeasonInsightsDataError",
     "PenaSeasonNotFoundError",
