+++ backend/src/persistence/application/use_cases/__init__.py
@@ -1,3 +1,6 @@
+from core.application.models import PenaInfo as PenaSummary
+from core.application.models import PenasPage
+from core.application.use_cases.get_penas_usecase import GetPenasUseCase
 from core.application.use_cases.get_season_match_insights_usecase import (
     GetSeasonMatchInsightsUseCase,
 )
@@ -14,8 +17,6 @@ from .get_pena_players_usecase import (
     PenaPlayerInfo,
     PenaPlayersPage,
 )
-from .get_penas_usecase import GetPenasUseCase, PenasPage
-from .get_penas_usecase import PenaInfo as PenaSummary
 from .get_player_profile_usecase import GetPlayerProfileUseCase, PenaInfo, PlayerProfile
 from .link_user_to_pena_usecase import (
     InvalidLinkTokenError,
