+++ backend/tests/test_penas_controller.py
@@ -16,12 +16,12 @@ from api.interface.controller.v1.model.request.penas_request import (
 )
 from auth.dependencies import require_user
 from auth.session import SessionData
+from core.application.models import PenaInfo, PenasPage
 from fastapi import HTTPException
 from persistence.application.use_cases.generate_pena_link_token_usecase import (
     PenaAccessDeniedError,
     PenaLinkToken,
 )
-from persistence.application.use_cases.get_penas_usecase import PenaInfo, PenasPage
 from persistence.application.use_cases.link_user_to_pena_usecase import (
     InvalidLinkTokenError,
     UserAlreadyLinkedError,
